Update nuget spec (still used old-school approach)
We get an error while trying to build nugets from VS (on Mac and Win):
error MSB4044: The “FilterAssemblies” task was not given a value
for the required parameter “DesignTimeBuild”.

Decided to keep using nuspec and scripts to pack/push packages.
As soon as we can use VS - the scripts should be removed.